Steering and shooting
Modified the receiver code so that the robot can now steer when the
nunchuck joystick is moved in different directions.
Added code to receive incoming shoot signal, the motor is scarily fast
though so it is not actually run.
Added a validation signal so that the receiver can tell if it's
connected to the transmitter.

Note, the brushless motor we were planning on using to shoot cannot
reverse. We will therefore have to either change our strategy or find a
different motor/esc.